+++ context/app/static/js/components/detailPage/MetadataTable/MetadataTable.tsx
@@ -5,7 +5,6 @@ import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
 import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
-import Stack from '@mui/material/Stack';
 
 import { StyledTableContainer, HeaderCell } from 'js/shared-styles/tables';
 import IconTooltipCell from 'js/shared-styles/tables/IconTooltipCell';
@@ -35,10 +34,11 @@ function MetadataTable({ tableRows = [] as MetadataTableRow[], columns = default
               <TableRow key={row.key}>
                 <IconTooltipCell tooltipTitle={row?.description}>{row.key}</IconTooltipCell>
                 <TableCell>
-                  <Stack direction="row">
-                    {row.value}
-                    {row.key.endsWith('age_value') && <DonorAgeTooltip donorAge={row.value} />}
-                  </Stack>
+                  {row.key.endsWith('age_value') ? (
+                    <DonorAgeTooltip donorAge={row.value}>{row.value}</DonorAgeTooltip>
+                  ) : (
+                    row.value
+                  )}
                 </TableCell>
               </TableRow>
             ))}