+++ context/app/static/js/components/detailPage/ProcessedData/HelperPanel/HelperPanel.tsx
@@ -15,16 +15,11 @@ import { SecondaryBackgroundTooltip } from 'js/shared-styles/tooltips';
 
 import { formatDate } from 'date-fns/format';
 import { HelperPanelPortal } from '../../DetailLayout/DetailLayout';
-import useProcessedDataStore from '../store';
 import StatusIcon from '../../StatusIcon';
-import { getDateLabelAndValue } from '../../utils';
+import { getDateLabelAndValue, useCurrentDataset } from '../../utils';
 import { HelperPanelButton } from './styles';
 import { useTrackEntityPageEvent } from '../../useTrackEntityPageEvent';
 import ProcessedDataWorkspaceMenu from '../ProcessedDataWorkspaceMenu';
-
-function useCurrentDataset() {
-  return useProcessedDataStore((state) => state.currentDataset);
-}
 
 function HelperPanelHeader() {
   const currentDataset = useCurrentDataset();
@@ -47,7 +42,7 @@ function HelperPanelStatus() {
   return (
     <Stack direction="row" alignItems="center">
       <StatusIcon status={currentDataset.status} />
-      <Typography variant="body2">{currentDataset.status}</Typography>
+      <Typography variant="body2">{`${currentDataset.status} (${currentDataset.mapped_data_access_level})`}</Typography>
     </Stack>
   );
 }
@@ -99,7 +94,7 @@ function HelperPanelActions() {
     return null;
   }
 
-  const { hubmap_id, uuid, status } = currentDataset;
+  const { hubmap_id, uuid, status, mapped_data_access_level } = currentDataset;
 
   return (
     <>
@@ -109,7 +104,7 @@ function HelperPanelActions() {
             <HelperPanelButton startIcon={<WorkspacesIcon color="primary" />}>Workspace</HelperPanelButton>
           </SecondaryBackgroundTooltip>
         }
-        datasetDetails={{ hubmap_id, uuid, status }}
+        datasetDetails={{ hubmap_id, uuid, status, mapped_data_access_level }}
         dialogType="ADD_DATASETS_FROM_HELPER_PANEL"
       />
       <SecondaryBackgroundTooltip title="Scroll down to the Bulk Data Transfer Section.">