+++ context/app/static/js/components/detailPage/visualization/VisualizationWorkspaceButton/VisualizationWorkspaceButton.tsx
@@ -7,30 +7,21 @@ import NewWorkspaceDialog from 'js/components/workspaces/NewWorkspaceDialog';
 import { useCreateWorkspaceForm } from 'js/components/workspaces/NewWorkspaceDialog/useCreateWorkspaceForm';
 import { WhiteBackgroundIconTooltipButton } from 'js/shared-styles/buttons';
 import { useAppContext } from 'js/components/Contexts';
+import { useCurrentDataset } from 'js/components/detailPage/utils';
 
 const tooltip = 'Launch New Workspace';
 
-interface VisualizationWorkspaceButtonProps {
-  uuid?: string;
-  hubmap_id?: string;
-  hasNotebook?: boolean;
-  mapped_data_access_level?: string;
-}
-
-function VisualizationWorkspaceButton({
-  uuid = '',
-  hubmap_id,
-  hasNotebook,
-  mapped_data_access_level,
-}: VisualizationWorkspaceButtonProps) {
+function VisualizationWorkspaceButton() {
+  const currentDataset = useCurrentDataset();
   const { isWorkspacesUser } = useAppContext();
+
   const { setDialogIsOpen, removeDatasets, ...rest } = useCreateWorkspaceForm({
-    defaultName: hubmap_id,
+    defaultName: currentDataset?.hubmap_id,
     defaultTemplate: 'visualization',
-    initialSelectedDatasets: [uuid],
+    initialSelectedDatasets: currentDataset ? [currentDataset.uuid] : [],
   });
 
-  if (!isWorkspacesUser ?? !uuid ?? !hubmap_id ?? !hasNotebook ?? mapped_data_access_level === 'Protected') {
+  if (!isWorkspacesUser || !currentDataset?.hubmap_id || currentDataset?.mapped_data_access_level !== 'Public') {
     return null;
   }