+++ context/app/static/js/pages/Dataset/utils.ts
@@ -1,11 +1,16 @@
 import { ProcessedDatasetInfo } from './hooks';
 
+const potentialDuplicateStates = ['error'];
+
 export function datasetSectionId(
-  dataset: Pick<ProcessedDatasetInfo, 'pipeline' | 'hubmap_id' | 'status'>,
+  dataset: Pick<ProcessedDatasetInfo, 'hubmap_id' | 'status'> & { pipeline?: string },
   prefix = '',
 ) {
   const { pipeline, hubmap_id, status } = dataset;
   const formattedDatasetIdentifier = (pipeline ?? hubmap_id).replace(/\s/g, '');
   const formattedStatus = status.replace(/\s/g, '');
-  return `${prefix}-${encodeURIComponent(formattedDatasetIdentifier)}-${encodeURIComponent(formattedStatus)}`.toLowerCase();
+  const deduplicatedFormattedStatus = potentialDuplicateStates.includes(formattedStatus.toLowerCase())
+    ? `${formattedStatus}-${hubmap_id}`
+    : formattedStatus;
+  return `${prefix}-${encodeURIComponent(formattedDatasetIdentifier)}-${encodeURIComponent(deduplicatedFormattedStatus)}`.toLowerCase();
 }