+++ context/app/static/js/shared-styles/tooltips/InfoTextTooltip/index.ts
@@ -1,0 +1,3 @@
+import InfoTextTooltip from './InfoTextTooltip';
+
+export default InfoTextTooltip;